PR feedback - Move Table within LoadBalancerLanding component
